refactor(plugin-sdk): share HTML escaping
