+++ src/plugin-sdk/text-utility-runtime.ts
@@ -2,6 +2,8 @@
 import type { BaseProbeResult } from "../channels/plugins/types.public.js";
 import { withTimeout } from "../utils/with-timeout.js";
 
+export { escapeHtml } from "../shared/html-escape.js";
+
 type ChannelProbeResult = BaseProbeResult & { elapsedMs?: number };
 
 /** Run a channel probe with shared timeout, elapsed-time, and error-result handling. */
@@ -27,16 +29,6 @@ export async function runChannelProbe<
     }
     return finish(onError(error));
   }
-}
-
-/** Escapes text for safe insertion into HTML text and quoted attribute values. */
-export function escapeHtml(value: string): string {
-  return value
-    .replaceAll("&", "&amp;")
-    .replaceAll("<", "&lt;")
-    .replaceAll(">", "&gt;")
-    .replaceAll('"', "&quot;")
-    .replaceAll("'", "&#39;");
 }
 
 export {
