+++ src/shared/html-escape.ts
@@ -1,0 +1,9 @@
+/** Escapes text for safe insertion into HTML text and quoted attribute values. */
+export function escapeHtml(value: string): string {
+  return value
+    .replaceAll("&", "&amp;")
+    .replaceAll("<", "&lt;")
+    .replaceAll(">", "&gt;")
+    .replaceAll('"', "&quot;")
+    .replaceAll("'", "&#39;");
+}
